Extract PredicateCollectionCondition from AllMatch, AnyMatch and NoneMatch
